requirements.txt
python-dotenv==1.0.1
newsdataapi==0.1.21
eventregistry==9.1
openai==1.54.0
fastapi==0.115.4
uvicorn==0.32.0
